Rename and clean up libraries; purge useless files
