--- classifiers/__init__.py
@@ -1,2 +1,0 @@
-from classifiers.k_nearest_neighbor import *
-from classifiers.linear_classifier import *
