+++ preprocessors/cumulative_moving_average.py
